VEOSS-1689 Updates for navigation and recurside delete function added
